Create STATION(TRANSMITTER)
ESP-NOW Transmitter with Button State
This project demonstrates how to use ESP-NOW communication protocol to transmit data from an ESP32 board to a receiver. The transmitter reads the state of a button and sends this state, along with a unique ID, to the receiver. The receiver can then use this information to perform actions such as lighting up an LED.